% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfgeom.R
\name{cfgeom}
\alias{cfgeom}
\title{Characteristic function of a geometric distribution}
\usage{
cfgeom(t, prob)
}
\arguments{
\item{t}{Values at which to evaluate the characteristic function}

\item{prop}{Probability of success of each trial. 0 < p <= 1}
}
\value{
A complex number
}
\description{
Characteristic function of a geometric distribution
}
