% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{.chypogeom}
\alias{.chypogeom}
\title{Compute a vector of CDF values of the hypogeometric distribution}
\usage{
.chypogeom(probs, ...)
}
\arguments{
\item{probs}{Vector of the probabilities of success of each geometric random
variable}

\item{...}{Other parameters to pass to .dhypogeom}
}
\value{
A vector of numerics of length N, where N is a truncation that ensures
the desired precision. The \code{k}th entry of the return vector is \verb{Pr(X < )}
}
\description{
Compute a vector of CDF values of the hypogeometric distribution
}
