% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{.dhypogeom}
\alias{.dhypogeom}
\title{Compute a vector of PMF values of the hypogeometric distribution}
\usage{
.dhypogeom(probs, eps = .Machine$double.eps, max_N = Inf)
}
\arguments{
\item{probs}{Vector of the probabilities of success of each geometric random
variable}

\item{eps}{Error tolerance for density}

\item{max_N}{Maximum N at which to truncate the densities}
}
\value{
A vector of numerics of length N, where N is a truncation that ensures
the desired precision. The \code{k}th entry of the return vector is \code{Pr(X = K-1)}
}
\description{
Compute a vector of PMF values of the hypogeometric distribution
}
