% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogeom.R
\name{Hypogeometric}
\alias{Hypogeometric}
\alias{dhypogeom}
\alias{phypogeom}
\alias{qhypogeom}
\alias{rhypogeom}
\title{The Hypogeometric distribution}
\usage{
dhypogeom(x, probs, log = FALSE, ...)

phypogeom(q, probs, lower.tail = TRUE, log.p = FALSE, ...)

qhypogeom(p, probs, lower.tail = TRUE, log.p = FALSE, ...)

rhypogeom(n, prob)
}
\arguments{
\item{x, q}{Vector of quantiles at which to evaluate the density or CDF}

\item{probs}{Vector of success probabilities of constituent geometric
distributions. Each entry must be in (0, 1]}

\item{log, log.p}{Logical; if \code{TRUE}, probabilities are given on the log scale}

\item{...}{Additional parameters to pass to the internal function\code{.dhypogeom}.}

\item{lower.tail}{Logical; if \code{TRUE} gives CDF otherwise gives hazard function}

\item{n}{Number of observations}
}
\value{
A vector of the same length as \code{x} or \code{q}
}
\description{
Density, distribution function, quantile function and random gemeration for a
hypogeometric distribution.
}
\details{
A hypogeometric random variable with parameter vector \code{probs} has the
distribution of a sum of independent geometric random variables with parameters
\code{probs[i]}.
}
